+++ CHANGELOG.md
@@ -1,8 +1,9 @@
-## [Unreleased](https://github.com/ScoopInstaller/Scoop/compare/v0.2.4...develop)
+## [Unreleased](https://github.com/ScoopInstaller/Scoop/compare/master...develop)
 
 ### Features
 
 - **scoop-(un)hold:** Support `scoop (un)hold scoop` ([#5089](https://github.com/ScoopInstaller/Scoop/issues/5089))
+- **scoop-update:** Stash uncommitted changes before update ([#5091](https://github.com/ScoopInstaller/Scoop/issues/5091))
 
 ## [v0.2.4](https://github.com/ScoopInstaller/Scoop/compare/v0.2.3...v0.2.4) - 2022-08-08